[feat] #149 sharedViewModel 확장함수 분리
