+++ presentation/src/main/java/com/mashup/presentation/common/extension/NavBackStackEntry.kt
@@ -1,0 +1,23 @@
+package com.mashup.presentation.common.extension
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.ViewModel
+import androidx.navigation.NavBackStackEntry
+import androidx.navigation.NavController
+
+/**
+ * Ssam_D_Android
+ * @author jaesung
+ * @created 2023/07/20
+ */
+@Composable
+inline fun <reified T : ViewModel> NavBackStackEntry.sharedViewModel(navController: NavController): T {
+    val navGraphRoute = destination.parent?.route ?: return hiltViewModel()
+    val parentEntry = remember(this) {
+        navController.getBackStackEntry(navGraphRoute)
+    }
+
+    return hiltViewModel(parentEntry)
+}
